Allow Cron jobs to fail and support adhoc runs.
I.e. `cron_schedule` is now optional so that the same job can be run as
a cron and then used for ad-hoc reporting outside of the scheduled job.
Cron jobs will now fail instead of aborting so that they can be
retried. However the failed job will not create a new scheduled
instance on completion since the scheduled instance is created
immediately on failure.
Remove Rocket Job prefix from metrics.